Migrate a bunch of half-coherent effects into Layers
